refactor: VacancyDetail decomposition
